+++ R/multiprocess.R
@@ -22,9 +22,24 @@
 #'
 #' @export
 multiprocess <- function(..., workers = availableCores(), envir = parent.frame()) {
-  .Deprecated(msg = sprintf("Detected creation of a 'multiprocess' future. Strategy 'multiprocess' is deprecated in future (>= 1.20.0) [2020-10-30]. Instead, explicitly specify either 'multisession' (recommended) or 'multicore'. Starting with future 1.31.0 [2023-01-31], 'multiprocess' is the same as 'sequential'."))
+
+  msg <- "Detected creation of a 'multiprocess' future, which are defunct in future (>= 1.32.0) [2023-03-06]."
+  msg <- paste(msg, "Instead, specify either 'multisession' (recommended) or 'multicore'.")
+  defunct <- getOption("future.deprecated.defunct")
+  if (is.element("multiprocess", defunct)) {
+    ## Need to wrap .Defunct() in another frame to avoid:
+    ## Error in as.vector(x, "character") : 
+    ##   cannot coerce type 'closure' to vector of type 'character'
+    dfcn <- function(...) .Defunct(...)
+  } else {
+    msg <- paste(msg, "If still used, 'multiprocess' becomes the same as 'sequential'.")
+    dfcn <- .Deprecated
+  }
+  dfcn(msg = msg, package = .packageName)
+  
   sequential(..., envir = envir)
 }
 class(multiprocess) <- c("sequential", "uniprocess", "future", "function")
 ## future (> 1.30.0): 'multiprocess' always resolves to 'sequential'
+class(multiprocess) <- c(class(multiprocess), "multiprocess")
 attr(multiprocess, "init") <- FALSE