+++ tests/future,optsenvvars.R
@@ -30,16 +30,38 @@ for (strategy in strategies) {
     identical(Sys.getenv("R_DEFAULT_INTERNET_TIMEOUT"), "300")
     #DISABLED# is.na(Sys.getenv("ABC", NA_character_))
   )
+  
+  envvars <- Sys.getenv()
   if (.Platform$OS.type == "windows") {
     ## Drop empty environment variables, because they are not supported by
     ## MS Windows, but may exist because they're inherited from a host OS
     old_envvars <- old_envvars[nzchar(old_envvars)]
-    envvars <- Sys.getenv()
     envvars <- envvars[nzchar(envvars)]
-    stopifnot(identical(envvars[names(old_envvars)], old_envvars))
-  } else {
-    stopifnot(identical(Sys.getenv()[names(old_envvars)], old_envvars))
   }
+
+  envvars <- envvars[names(old_envvars)]
+  
+  ## Any added?
+  diff <- setdiff(names(envvars), names(old_envvars))
+  if (length(diff) > 0) {
+    stop(sprintf("Detected added environment variables: %s", paste(diff, collapse = ", ")))
+  }
+  
+  ## Any removed?
+  diff <- setdiff(names(old_envvars), names(envvars))
+  if (length(diff) > 0) {
+    stop(sprintf("Detected removed environment variables: %s", paste(diff, collapse = ", ")))
+  }
+
+  ## Any changed?
+  for (name in names(old_envvars)) {
+      if (!identical(envvars[[name]], old_envvars[[name]])) {
+          stop(sprintf("Detected modified environment variable: %s=%s (was %s)", name, sQuote(envvars[[name]]), sQuote(old_envvars[[name]])))
+      }
+  }
+
+  stopifnot(all.equal(envvars, old_envvars))
+  stopifnot(identical(envvars, old_envvars))
 
   message(sprintf("- plan('%s') ... DONE", strategy))
 } ## for (strategy ...)